implement simple book service and unit tests

+++ src/main/java/com/krzywdek19/readingTacker/book/Book.java
@@ -21,6 +21,6 @@ public class Book {
     @ManyToOne
     @JoinColumn(name = "author_id")
     private Author author;
-    private Long pages;
+    private Integer pages;
     private String description;
 }
